requirements.txt
aiofiles==0.8.0
aiohttp==3.8.1
aiosignal==1.2.0
anyio==3.6.1
async-timeout==4.0.2
attrs==22.1.0
beautifulsoup4==4.11.1
bs4==0.0.1
cffi==1.15.1
charset-normalizer==2.1.1
Deprecated==1.2.13
discord.py==2.0.1
frozenlist==1.3.1
h11==0.12.0
httpcore==0.15.0
httpx==0.23.0
idna==3.3
multidict==6.0.2
packaging==21.3
pycparser==2.21
PyNaCl==1.5.0
pyparsing==3.0.9
redis==4.3.4
requests==2.28.1
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.2.post1
spotipy==2.20.0
typing_extensions==4.3.0
urllib3==1.26.12
wincertstore==0.2
wrapt==1.14.1
yarl==1.8.1
youtube-dl==2021.12.17
